Refactor game package and add improve the game loop design.
