+++ game/src/main/kotlin/xlitekt/game/content/command/CommandListener.kt
@@ -1,4 +1,4 @@
-package xlitekt.game.command
+package xlitekt.game.content.command
 
 import xlitekt.game.actor.player.Player
