+++ game/src/main/kotlin/xlitekt/game/content/container/Container.kt
@@ -1,3 +1,3 @@
-package xlitekt.game.container
+package xlitekt.game.content.container
 
 data class Container(val id: Int, val size: Int)
