+++ game/src/main/kotlin/xlitekt/game/content/friend/Friend.kt
@@ -1,3 +1,3 @@
-package xlitekt.game.friend
+package xlitekt.game.content.friend
 
 data class Friend(val displayName: String, val visible: Boolean)
