+++ game/src/main/kotlin/xlitekt/game/content/vars/VarType.kt
@@ -1,4 +1,4 @@
-package xlitekt.game.vars
+package xlitekt.game.content.vars
 
 /**
  * @author Jordan Abraham
