+++ game/src/main/kotlin/xlitekt/game/content/vars/VarsMap.kt
@@ -1,4 +1,4 @@
-package xlitekt.game.vars
+package xlitekt.game.content.vars
 
 import xlitekt.shared.inject
 import xlitekt.shared.resource.VarBits
