--- game/src/main/kotlin/xlitekt/game/world/engine/Task.kt
@@ -1,9 +1,0 @@
-package xlitekt.game.world.engine
-
-/**
- * @author Jordan Abraham
- */
-interface Task {
-    fun start()
-    fun shutdown()
-}
