+++ script/src/main/kotlin/script/packet/disassembler/handler/ClientCheatHandler.kts
@@ -1,6 +1,6 @@
 package script.packet.disassembler.handler
 
-import xlitekt.game.command.Commands
+import xlitekt.game.content.command.Commands
 import xlitekt.game.packet.ClientCheatPacket
 import xlitekt.game.packet.disassembler.handler.onPacketHandler
