+++ script/src/main/kotlin/script/ui/AdvancedSettingsListener.kts
@@ -1,12 +1,12 @@
 package script.ui
 
-import xlitekt.game.ui.InterfaceEvent
-import xlitekt.game.ui.UserInterface
-import xlitekt.game.ui.UserInterfaceEvent
-import xlitekt.game.ui.onInterface
-import xlitekt.game.vars.VarBit
-import xlitekt.game.vars.VarPlayer
-import xlitekt.game.vars.Vars
+import xlitekt.game.content.ui.InterfaceEvent
+import xlitekt.game.content.ui.UserInterface
+import xlitekt.game.content.ui.UserInterfaceEvent
+import xlitekt.game.content.ui.onInterface
+import xlitekt.game.content.vars.VarBit
+import xlitekt.game.content.vars.VarPlayer
+import xlitekt.game.content.vars.Vars
 
 private val closeSettingsChildId = 4
 private val firstLayerChildId = 19
