+++ script/src/main/kotlin/script/ui/EquipmentStatsListener.kts
@@ -1,8 +1,8 @@
 package script.ui
 
 import xlitekt.game.actor.player.script
-import xlitekt.game.ui.UserInterface
-import xlitekt.game.ui.onInterface
+import xlitekt.game.content.ui.UserInterface
+import xlitekt.game.content.ui.onInterface
 
 onInterface<UserInterface.EquipmentStats> {
     onCreate {
