+++ script/src/main/kotlin/script/ui/MiniMapListener.kts
@@ -1,8 +1,8 @@
 package script.ui
 
-import xlitekt.game.ui.UserInterface
-import xlitekt.game.ui.onInterface
-import xlitekt.game.vars.VarPlayer
+import xlitekt.game.content.ui.UserInterface
+import xlitekt.game.content.ui.onInterface
+import xlitekt.game.content.vars.VarPlayer
 
 onInterface<UserInterface.MiniMap> {
     onClick("Toggle Run") {
