+++ script/src/main/kotlin/script/ui/WornEquipmentListener.kts
@@ -2,9 +2,9 @@ package script.ui
 
 import xlitekt.game.actor.player.Equipment.Companion.SLOT_WEAPON
 import xlitekt.game.actor.player.message
-import xlitekt.game.item.Item
-import xlitekt.game.ui.UserInterface
-import xlitekt.game.ui.onInterface
+import xlitekt.game.content.item.Item
+import xlitekt.game.content.ui.UserInterface
+import xlitekt.game.content.ui.onInterface
 
 private val equipmentContainerKey = 94
 private val equipmentSize = 15
